Revert "final"
This reverts commit 12d89b699d622199124102854118de13cf4f43dd.

+++ Game_Server/Card.cpp
@@ -1,6 +1,6 @@
 #include "Card.h"
 
-Card::Card(byte type, byte num, byte score): m_Card_type(type),m_Card_num(num),m_Card_Score(score)
+Card::Card(int type, int num, int score): m_Card_type(type),m_Card_num(num),m_Card_Score(score)
 {
 }
